up_and_running/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
click==8.1.7
distro==1.9.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
openai==1.13.3
psycopg2-binary==2.9.9
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
python-dotenv==1.0.1
rich==13.7.1
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.10.0
